move test to a single node test
